+++ src/main/java/ru/hits/attackdefenceplatform/public_interface/team/TeamListDto.java
@@ -7,7 +7,7 @@ public record TeamListDto(
         String name,
         Integer place,
         Double points,
-        Long userCount,
+        Integer userCount,
         Long membersCount,
         Boolean isMyTeam,
         String ipAddress
